Update http.rb adapter
