+++ lib/webmock.rb
@@ -40,7 +40,7 @@ require 'webmock/api'
 require 'webmock/http_lib_adapters/http_lib_adapter_registry'
 require 'webmock/http_lib_adapters/http_lib_adapter'
 require 'webmock/http_lib_adapters/net_http'
-require 'webmock/http_lib_adapters/http_gem_adapter'
+require 'webmock/http_lib_adapters/http_rb_adapter'
 require 'webmock/http_lib_adapters/httpclient_adapter'
 require 'webmock/http_lib_adapters/patron_adapter'
 require 'webmock/http_lib_adapters/curb_adapter'
