--- lib/webmock/http_lib_adapters/http_gem/streamer.rb
@@ -1,13 +1,0 @@
-module HTTP
-  class Response
-    class Streamer
-      def initialize(str)
-        @io = StringIO.new str
-      end
-
-      def readpartial(size = HTTP::Client::BUFFER_SIZE)
-        @io.read size
-      end
-    end
-  end
-end
