+++ lib/webmock/http_lib_adapters/http_rb/response.rb
@@ -3,7 +3,7 @@ module HTTP
     def to_webmock
       webmock_response = ::WebMock::Response.new
 
-      webmock_response.status  = [status, reason]
+      webmock_response.status  = [status.to_i, reason]
       webmock_response.body    = body.to_s
       webmock_response.headers = headers.to_h
 
@@ -11,7 +11,7 @@ module HTTP
     end
 
     def self.from_webmock(webmock_response, request_signature = nil)
-      status  = webmock_response.status.first
+      status  = Status.new(webmock_response.status.first)
       headers = webmock_response.headers || {}
       body    = Body.new Streamer.new webmock_response.body
       uri     = URI request_signature.uri.to_s if request_signature
